Move evaluation.py into parent folder
